requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
redis==4.6.0
celery==5.3.1
flower==2.0.1
pandas==2.1.4
numpy==1.26.4
openpyxl==3.1.2
reportlab==4.0.4
Pillow==11.3.0
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.2.1
blinker==1.9.0
SQLAlchemy==2.0.42
greenlet==3.2.3
prometheus-client==0.22.1 
flask-caching